both.go: refactors for merge functionality
This finishes (at least for now) the refactor for the merge struct and
the surrounding code of the old cli.go:run() routine. The different
large sections of nesting have been pulled out into functions and most
of the mutable state has been encapsulated in the struct, and many
things were made immutable in the process.

It could be a lot better, but this is a good start.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>